% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{binary_SeSp}
\alias{binary_SeSp}
\title{Simulate A Binary Outcome}
\usage{
binary_SeSp(prevalence, n, Se, Sp, reshuffle = TRUE, ...)
}
\arguments{
\item{prevalence}{observed prevalence of the event (coded as 1) in
the population.}

\item{n}{population size.}

\item{Se}{sensitivity of the classifier.}

\item{Sp}{specificity of the classifier.}

\item{reshuffle}{logical, should the output be randomly re-shuffled?
defaults to \code{TRUE}.}

\item{...}{extra arguments, currently none.}
}
\value{
an object of class \code{simOutcome}. Technically, data frame with
two columns \code{outcome} and \code{predicted} and a \code{summary()} method returning
statistics such as accuracy, sensitivity, specificity, ROC AUC
(are under the receiver-operator characteristic curve) and Cohen's kappa.
}
\description{
Simulates an outcome of a binary classifier (0/1) with the pre-defined
prevalence of the event (coded as 1), the sensitivity and specificity.
}
\details{
The total number of positive cases, true positives and true negatives
computed as defined by \code{n}, \code{prevalence}, \code{Se} and \code{Sp}
are each rounded to the nearest integer.
}
