% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binary_stats}
\alias{binary_stats}
\title{Compute Binary Model Stats.}
\usage{
binary_stats(data)
}
\arguments{
\item{data}{a data frame with the columns \code{outcome} and \code{predicted}.
The event or positivity is coded as 1.}
}
\value{
a data frame with the following columns:
\itemize{
\item \code{n}: the total number of observations
\item \code{n_events}: the number of observed events coded in the
input data frame as 1
\item \code{n_predicted}: the number of predicted events
\item \code{prev_events}: the observed prevalence
\item \code{prev_predicted}: the predicted prevalence
\item \code{accuracy}: overall accuracy
\item \code{kappa}: Cohen's kappa
\item \code{Se}: sensitivity
\item \code{Sp}: specificity
\item \code{AUC}: are under the ROC curve
}
}
\description{
Computes fit statistics for a simulated binary classifier.
}
\details{
Technically it employs the \code{caret} package functions
\code{\link[caret]{defaultSummary}} and \code{\link[caret]{twoClassSummary}}.
}
