% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{summary.simOutcome}
\alias{summary.simOutcome}
\title{Model Statistic Summary for the simOutcome Objects.}
\usage{
\method{summary}{simOutcome}(object, ...)
}
\arguments{
\item{object}{an object of the \code{simOutcome} class.}

\item{...}{extra arguments, currently none.}
}
\description{
Computes model fit stats for the \code{simOutcome} objects.
}
\details{
The fit stats are model type-specific.
See \code{\link{binary_stats}} for simulated binary classifiers.
}
